requirements.txt
sqlalchemy==2.0.41
pydantic==2.11.7
psycopg2==2.9.10
google-cloud-bigquery==3.35.0